Added real mocks with tests for OpenTable/Detail, running into problems with int vs float types

+++ tests/PilotLive/Models/ItemModelTest.php
@@ -27,7 +27,7 @@ class ItemModelTest extends BaseModelTest
                 'max' => 1000000.00,
             ],
             'qty' => [
-                'type' => 'double',
+                'type' => 'integer',
                 'nullable' => false,
                 'readonly' => false,
                 'min' => 0.00,
